refactor - Ladder -> LadderDto , LadderRows -> Ladder 로 변경
