+++ src/main/java/nextstep/ladder/domain/Ladder.java
@@ -1,20 +1,22 @@
 package nextstep.ladder.domain;
 
+import java.util.Collections;
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
+
 public class Ladder {
 
-    private final Users users;
-    private final LadderRows ladderRows;
+    private List<LadderRow> ladderRows;
 
     public Ladder(Users users, LadderHeight height) {
-        this.users = users;
-        this.ladderRows = new LadderRows(users, height);
+        this.ladderRows = Collections.unmodifiableList(IntStream.range(0, height.getValue())
+                .boxed()
+                .map(index -> new LadderRow(users.size()))
+                .collect(Collectors.toList()));
     }
 
-    public Users getUsers() {
-        return users;
-    }
-
-    public LadderRows getLadderRows() {
-        return ladderRows;
+    public List<LadderRow> getLadderRows() {
+        return this.ladderRows;
     }
 }
